Use int arrays.

+++ src/main/java/dev/morling/onebrc/CalculateAverage_ericxiao.java
@@ -32,7 +32,7 @@ public class CalculateAverage_ericxiao {
 
     private static final String FILE = "./measurements.txt";
 
-    static class ProcessFileMap implements Callable<Map<ProcessFileMap.KeySlice, double[]>> {
+    static class ProcessFileMap implements Callable<Map<ProcessFileMap.KeySlice, int[]>> {
         private long readStart;
         private long readEnd;
         private boolean lastRead;
@@ -48,7 +48,7 @@ public class CalculateAverage_ericxiao {
             this.firstRead = firstRead;
         }
 
-        private final HashMap<KeySlice, double[]> hashMap = new HashMap<>();
+        private final HashMap<KeySlice, int[]> hashMap = new HashMap<>();
 
         private static Unsafe initUnsafe() {
             try {
@@ -109,8 +109,8 @@ public class CalculateAverage_ericxiao {
 
             final byte negativeSign = '-';
             final byte periodSign = '.';
-            double accumulator = 0;
-            double multiplier = 1;
+            int accumulator = 0;
+            int multiplier = 1;
 
             for (int i = keyLength + 1; i <= keyLength + valueLength; ++i) {
                 if (entryBytes[i] == negativeSign) {
@@ -119,12 +119,12 @@ public class CalculateAverage_ericxiao {
                 else if (entryBytes[i] != periodSign)
                     accumulator = accumulator * 10 + entryBytes[i] - '0';
             }
-            double value = multiplier == -1 ? -accumulator : accumulator;
+            int value = multiplier == -1 ? -accumulator : accumulator;
 
             hashMap.compute(key, (k, v) -> {
                 if (v == null) {
                     k.materialize();
-                    return new double[]{ value, value, value, 1 };
+                    return new int[]{ value, value, value, 1 };
                 }
                 else {
                     v[0] = Math.min(v[0], value);
@@ -141,11 +141,11 @@ public class CalculateAverage_ericxiao {
             return (mask - 0x0101010101010101L) & (~mask & 0x8080808080808080L);
         }
 
-        public Map<KeySlice, double[]> call() {
+        public Map<KeySlice, int[]> call() {
             return readMemory(readStart, readEnd);
         }
 
-        private Map<KeySlice, double[]> readMemory(long startAddress, long endAddress) {
+        private Map<KeySlice, int[]> readMemory(long startAddress, long endAddress) {
             int packedBytes = 0;
             final long singleSemiColonPattern = 0x3BL;
             final long semiColonPattern = 0x3B3B3B3B3B3B3B3BL;
@@ -257,7 +257,7 @@ public class CalculateAverage_ericxiao {
     public static void main(String[] args) throws Exception {
         int numThreads = Runtime.getRuntime().availableProcessors() - 1; // Use the number of available processors
         ExecutorService executorService = Executors.newFixedThreadPool(numThreads);
-        List<Callable<Map<ProcessFileMap.KeySlice, double[]>>> callableTasks = new ArrayList<>();
+        List<Callable<Map<ProcessFileMap.KeySlice, int[]>>> callableTasks = new ArrayList<>();
         Path filePath = Path.of(FILE);
 
         try (FileChannel fileChannel = (FileChannel) Files.newByteChannel(filePath, EnumSet.of(StandardOpenOption.READ))) {
@@ -278,10 +278,10 @@ public class CalculateAverage_ericxiao {
 
             callableTasks.add(new ProcessFileMap(readStart, readStart + readLength, false, true));
 
-            List<Map<ProcessFileMap.KeySlice, double[]>> results = new ArrayList<>();
+            List<Map<ProcessFileMap.KeySlice, int[]>> results = new ArrayList<>();
             try {
-                List<Future<Map<ProcessFileMap.KeySlice, double[]>>> futures = executorService.invokeAll(callableTasks);
-                for (Future<Map<ProcessFileMap.KeySlice, double[]>> future : futures) {
+                List<Future<Map<ProcessFileMap.KeySlice, int[]>>> futures = executorService.invokeAll(callableTasks);
+                for (Future<Map<ProcessFileMap.KeySlice, int[]>> future : futures) {
                     try {
                         results.add(future.get());
                     }
@@ -296,11 +296,11 @@ public class CalculateAverage_ericxiao {
             finally {
                 executorService.shutdown();
                 // fileChannel.close();
-                Map<ProcessFileMap.KeySlice, double[]> mapA = results.getFirst();
+                Map<ProcessFileMap.KeySlice, int[]> mapA = results.getFirst();
                 for (int i = 1; i < numThreads; ++i) {
                     results.get(i).forEach((station, stationMeasurements) -> {
                         if (mapA.containsKey(station)) {
-                            double[] measurements = mapA.get(station);
+                            int[] measurements = mapA.get(station);
                             measurements[0] = Math.min(measurements[0], stationMeasurements[0]);
                             measurements[1] = Math.max(measurements[1], stationMeasurements[1]);
                             measurements[2] = measurements[2] + stationMeasurements[2];
@@ -314,9 +314,9 @@ public class CalculateAverage_ericxiao {
                 // print key and values
                 int counter = 1;
                 System.out.print("{");
-                for (Map.Entry<ProcessFileMap.KeySlice, double[]> entry : mapA.entrySet()) {
-                    double[] measurements = entry.getValue();
-                    double mean = measurements[2] / measurements[3];
+                for (Map.Entry<ProcessFileMap.KeySlice, int[]> entry : mapA.entrySet()) {
+                    int[] measurements = entry.getValue();
+                    double mean = (double) measurements[2] / (double) measurements[3];
                     System.out.print(entry.getKey().key + "=" + (measurements[0] / 10.0) + "/"
                             + (Math.round(mean) / 10.0) + "/"
                             + (measurements[1]) / 10.0);
